requirements.txt
beautifulsoup4==4.13.2
requests==2.32.4
pydantic==2.11.2
redis==5.0.0
python-dotenv==1.1.1
fastapi==0.115.0
uvicorn==0.30.6
uvloop==0.22.1
httptools==0.7.1
watchfiles==1.1.1
websockets==12.0
pymongo==4.6.1
motor==3.3.2
playwright==1.51.0
apscheduler==3.10.4
ruff==0.4.4
